feat: 커스텀 구분자 분리 구현 커스텀 구분자가 지정되어있음에도 다른 특수문자 포함된 경우 예외처리
